+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-11T13:44:33.441Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-11T15:08:58.855Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")