+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRBMH1CVG6CEMA8BB99SMPKG"
+        "01KRBSBMVN0NYC68ZBEZGPTD8V"
     )
     r"""The unique identifier of the data source"""