+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRBMH1CXYRRD5YEZHP7PS319"
+        "01KRBSBMVQRR3YHANT6R675Q4Z"
     )
     r"""The unique identifier of the data source"""