+++ packages/orq-rc/RELEASES.md
@@ -7749,3 +7749,13 @@ Based on:
 - [python v4.9.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.42 - packages/orq-rc
+
+## 2026-05-11 15:12:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.43 - packages/orq-rc